+++ lib/onlyoffice_digitalocean_wrapper/digitalocean_wrapper/getters.rb
@@ -22,9 +22,33 @@ module OnlyofficeDigitaloceanWrapper
       end
     end
 
+    # Get project by name
+    # @param [String] project_name
+    # @return [DropletKit::Project] object representing a project
+    def project_by_name(project_name)
+      retry_exception do
+        projects = @client.projects.all
+        projects.find { |x| x.name == project_name }
+      end
+    end
+
+    # Get project id by name
+    # @param [String] project_name
+    # @return [nil, String] id of current project or nil is no project found
+    def get_project_id_by_name(project_name)
+      project = project_by_name(project_name)
+      if project.nil?
+        logger.info("get_project_id_by_name(#{project_name}): not found any projects")
+        nil
+      else
+        logger.info("get_project_id_by_name(#{project_name}): #{project.id}")
+        project.id
+      end
+    end
+
     # Return droplet id by it's name
     # @param droplet_name [String] name of droplet
-    # @return [Integer] id of droplet
+    # @return [nil, Integer] id of droplet or nil is no one droplet found
     def get_droplet_id_by_name(droplet_name)
       droplet = droplet_by_name(droplet_name)
       if droplet.nil?